style: add project class to diff. file
